fix(sim): close persisted evidence validation gaps
